finish the api, cleanup, add test
