+++ cli/build.rs
@@ -331,6 +331,7 @@ deno_core::extension!(
   esm = [
     dir "js",
     "40_testing.js",
+    "40_jupyter.js",
     "99_main.js"
   ],
   customizer = |ext: &mut deno_core::Extension| {
