+++ runtime/js/90_deno_ns.js
@@ -2,6 +2,12 @@
 
 const core = globalThis.Deno.core;
 const ops = core.ops;
+const primordials = globalThis.__bootstrap.primordials;
+const {
+  Error,
+  ObjectDefineProperty,
+} = primordials;
+
 import * as timers from "ext:deno_web/02_timers.js";
 import * as httpClient from "ext:deno_fetch/22_http_client.js";
 import * as console from "ext:deno_console/01_console.js";
@@ -178,11 +184,14 @@ const denoNsUnstable = {
   Kv: kv.Kv,
   KvU64: kv.KvU64,
   KvListIterator: kv.KvListIterator,
-  jupyter: {
-    async sendIo(msg_type, content) {
-      await core.opAsync("op_jupyter_send_io", msg_type, content);
-    },
-  },
 };
+
+ObjectDefineProperty(denoNsUnstable, "jupyter", {
+  get() {
+    throw new Error(
+      "Deno.jupyter is only available in `deno jupyter` subcommand.",
+    );
+  },
+});
 
 export { denoNs, denoNsUnstable };
